Wrapping scene in Observable Object, SceneStore
The app will build and run, but...

The good news is that the GameScene is not get re-called over and over again, so our level stays in place as long as the view is alive.

Also, we can change values directly (e.g., scene.isPaused = true), but we can’t pass these around as bindings, e.g., so that another view like the modal can un-pause them.

The bad news is we don’t have any value bindings when values change in the scene.